Add comments to new rust fxa code, split app delegate up for clarity (#6280)
* Add comments to new rust fxa code

* Split up app delegate for push notification handling, and sync sent tab handling

* fix swift lint issues

* address nits
