+++ RustFxA/RustFirefoxAccounts.swift
@@ -8,6 +8,11 @@ import SwiftKeychainWrapper
 
 fileprivate let prefs = NSUserDefaultsPrefs(prefix: "profile")
 
+/**
+ A singleton that wraps the Rust FxA library.
+ The singleton design is poor for testability through dependency injection and may need to be changed in future.
+ */
+// TODO: renamed FirefoxAccounts.swift once the old code is removed fully.
 open class RustFirefoxAccounts {
     public static let prefKeyLastDeviceName = "prefKeyLastDeviceName"
 
@@ -15,12 +20,13 @@ open class RustFirefoxAccounts {
     public let redirectURL = "urn:ietf:wg:oauth:2.0:oob:oauth-redirect-webchannel"
     public static var shared = RustFirefoxAccounts()
     public let accountManager: FxAccountManager
-    public var avatar: Avatar? = nil
+    public var avatar: Avatar?
     private static var startupCalled = false
     public let syncAuthState: SyncAuthState
 
     public let pushNotifications = PushNotificationSetup()
-    
+
+    // This is used so that if a migration failed, show a UI indicator for the user to manually log in to their account.
     public var accountMigrationFailed: Bool {
         get {
             return UserDefaults.standard.bool(forKey: "fxaccount-migration-failed")
@@ -30,6 +36,15 @@ open class RustFirefoxAccounts {
         }
     }
 
+    /** Must be called before this class is fully usable. Until this function is complete,
+     all methods in this class will behave as if there is no Fx account.
+     It will be called on app startup, and extensions must call this before using the class.
+     If it is possible code could access `shared` before initialize() is complete, these callers should also
+     hook into notifications like `.accountProfileUpdate` to refresh once initialize() is complete.
+
+     The alternative implemention would be to have `shared` as a Deferred<RustFirefoxAccounts>. However that
+     would require a significant rewrite of existing code, for minimal added benefit.
+     */
     public static func startup(completion: ((RustFirefoxAccounts) -> Void)? = nil) {
         if startupCalled {
             completion?(shared)
@@ -39,15 +54,14 @@ open class RustFirefoxAccounts {
 
         shared.accountManager.initialize() { result in
             let hasAttemptedMigration = UserDefaults.standard.bool(forKey: "hasAttemptedMigration")
+
+            // Note this checks if startup() is called in an app extensions, and if so, do not try account migration
             if Bundle.main.bundleURL.pathExtension != "appex", let tokens = migrationTokens(), !hasAttemptedMigration {
                 UserDefaults.standard.set(true, forKey: "hasAttemptedMigration")
 
-                ["bookmarks", "history", "passwords", "tabs"].forEach {
-                    if let val = prefs.boolForKey("sync.engine.\($0).enabled"), !val {
-                        // is disabled
-                    }
-                }
-
+                // The client app only needs to trigger this one time. If it fails due to offline state, the rust library
+                // will automatically re-try until success or permanent failure (notifications accountAuthenticated / accountMigrationFailed respectively).
+                // See also `init()` use of `.accountAuthenticated` below.
                 shared.accountManager.authenticateViaMigration(sessionToken: tokens.session, kSync: tokens.ksync, kXCS: tokens.kxcs) { _ in }
             }
 
@@ -73,8 +87,7 @@ open class RustFirefoxAccounts {
 
         let config = FxAConfig(server: server, clientId: clientID, redirectUri: redirectURL)
         let type = UIDevice.current.userInterfaceIdiom == .pad ? DeviceType.tablet : DeviceType.mobile
-        let deviceConfig = DeviceConfig(name:  DeviceInfo.defaultClientName(), type: type, capabilities: [.sendTab])
-
+        let deviceConfig = DeviceConfig(name: DeviceInfo.defaultClientName(), type: type, capabilities: [.sendTab])
         let accessGroupPrefix = Bundle.main.object(forInfoDictionaryKey: "MozDevelopmentTeam") as! String
         let accessGroupIdentifier = AppInfo.keychainAccessGroupWithPrefix(accessGroupPrefix)
 
@@ -85,7 +98,9 @@ open class RustFirefoxAccounts {
                                             withLabel: RustFirefoxAccounts.syncAuthStateUniqueId,
                 factory: syncAuthStateCachefromJSON))
 
-        NotificationCenter.default.addObserver(forName: .accountAuthenticated,  object: nil, queue: .main) { [weak self] notification in
+        // Called when account is logged in for the first time, on every app start when the account is found (even if offline), and when migration of an account is completed.
+        NotificationCenter.default.addObserver(forName: .accountAuthenticated, object: nil, queue: .main) { [weak self] notification in
+            // Handle account migration completed successfully. Need to clear the old stored apnsToken and re-register push.
             if let type = notification.userInfo?["authType"] as? FxaAuthType, case .migrated = type {
                 KeychainWrapper.sharedAppContainerKeychain.removeObject(forKey: "apnsToken", withAccessibility: .afterFirstUnlock)
                 NotificationCenter.default.post(name: .RegisterForPushNotifications, object: nil)
@@ -94,7 +109,7 @@ open class RustFirefoxAccounts {
             self?.update()
         }
         
-        NotificationCenter.default.addObserver(forName: .accountProfileUpdate,  object: nil, queue: .main) { [weak self] notification in
+        NotificationCenter.default.addObserver(forName: .accountProfileUpdate, object: nil, queue: .main) { [weak self] notification in
             self?.update()
         }
 
@@ -109,6 +124,7 @@ open class RustFirefoxAccounts {
         }
     }
 
+    /// When migrating to new rust FxA, grab the old session tokens and try to re-use them.
     private class func migrationTokens() -> (session: String, ksync: String, kxcs: String)? {
         // Keychain forKey("profile.account"), return dictionary, from there
         // forKey("account.state.<guid>"), guid is dictionary["stateKeyLabel"]
@@ -137,29 +153,36 @@ open class RustFirefoxAccounts {
         return (session: sessionToken, ksync: ksync, kxcs: kxcs)
     }
 
+    /// This is typically used to add a UI indicator that FxA needs attention (usually re-login manually).
     public var isActionNeeded: Bool {
         if accountManager.accountMigrationInFlight() || accountMigrationFailed { return true }
         if !accountManager.hasAccount() { return false }
         return accountManager.accountNeedsReauth()
     }
 
+    /// Rust FxA notification handlers can call this to update caches and the UI.
     private func update() {
         let avatarUrl = accountManager.accountProfile()?.avatar?.url
         if let str = avatarUrl, let url = URL(string: str) {
             avatar = Avatar(url: url)
         }
+
+        // The userProfile (email, display name, etc) and the device name need to be cached for when the app starts in an offline state. Now is a good time to update those caches.
+
         // Accessing the profile will trigger a cache update if needed
-        let _ = userProfile
+        _ = userProfile
 
         // Update the device name cache
         if let deviceName = accountManager.deviceConstellation()?.state()?.localDevice?.displayName {
             UserDefaults.standard.set(deviceName, forKey: RustFirefoxAccounts.prefKeyLastDeviceName)
         }
-        
+
+        // The legacy system had both of these notifications for UI updates. Possibly they could be made into a single notification
         NotificationCenter.default.post(name: .FirefoxAccountProfileChanged, object: self)
         NotificationCenter.default.post(name: .FirefoxAccountStateChange, object: self)
     }
 
+    /// Cache the user profile (i.e. email, user name) for when the app starts offline. Notice this gets cleared when an account is disconnected.
     private let prefKeyCachedUserProfile = "prefKeyCachedUserProfile"
     private var cachedUserProfile: FxAUserProfile?
     public var userProfile: FxAUserProfile? {
@@ -190,6 +213,10 @@ open class RustFirefoxAccounts {
     }
 }
 
+/**
+ Wrap MozillaAppServices.Profile in an easy-to-serialize (and cache) FxAUserProfile.
+ Caching of this is required for when the app starts offline.
+ */
 public struct FxAUserProfile: Codable, Equatable {
     public let uid: String
     public let email: String
